Add WebexHelper class
